Added confirmation dialog before accidentally the whole repo

+++ init.sh
@@ -2,6 +2,13 @@
 
 set -e
 
+
+read -p "Are you sure? " -n 1
+if [[ ! $REPLY =~ ^[Yy]$ ]]
+then
+    exit 1
+fi
+
 rm -rf .git
 git init
